blacken

+++ activitysim/abm/tables/time_windows.py
@@ -16,10 +16,10 @@ logger = logging.getLogger(__name__)
 @inject.injectable(cache=True)
 def tdd_alts():
     # right now this file just contains the start and end hour
-    file_path = config.config_file_path('tour_departure_and_duration_alternatives.csv')
+    file_path = config.config_file_path("tour_departure_and_duration_alternatives.csv")
     df = pd.read_csv(file_path)
 
-    df['duration'] = df.end - df.start
+    df["duration"] = df.end - df.start
 
     # - NARROW
     df = df.astype(np.int8)
@@ -37,15 +37,17 @@ def tdd_alt_segments():
     # school,PM,15,17
     # school,EV,18,22
 
-    file_path = config.config_file_path('tour_departure_and_duration_segments.csv', mandatory=False)
+    file_path = config.config_file_path(
+        "tour_departure_and_duration_segments.csv", mandatory=False
+    )
 
     if file_path:
 
-        df = pd.read_csv(file_path, comment='#')
+        df = pd.read_csv(file_path, comment="#")
 
         # - NARROW
-        df['start'] = df['start'].astype(np.int8)
-        df['end'] = df['end'].astype(np.int8)
+        df["start"] = df["start"].astype(np.int8)
+        df["end"] = df["end"].astype(np.int8)
 
     else:
         df = None
@@ -58,7 +60,7 @@ def person_windows(persons, tdd_alts):
 
     df = tt.create_timetable_windows(persons, tdd_alts)
 
-    inject.add_table('person_windows', df)
+    inject.add_table("person_windows", df)
 
     return df
 
@@ -66,5 +68,5 @@ def person_windows(persons, tdd_alts):
 @inject.injectable()
 def timetable(person_windows, tdd_alts):
 
-    logging.debug('@inject timetable')
+    logging.debug("@inject timetable")
     return tt.TimeTable(person_windows.to_frame(), tdd_alts, person_windows.name)
